+++ sdk/storagediscovery/azure-mgmt-storagediscovery/generated_samples/storage_discovery_workspaces_update.py
@@ -39,7 +39,7 @@ def main():
                 "capabilities": {
                     "azureBlobStorage": {
                         "capacityDetails": {"status": "Enabled"},
-                        "prefixDefinitions": [
+                        "prefixConfigurations": [
                             {
                                 "containerName": "samplecontainer",
                                 "prefix": "sampleDir",