+++ sdk/storagediscovery/azure-mgmt-storagediscovery/generated_tests/test_storage_discovery_mgmt_storage_discovery_workspaces_operations.py
@@ -47,7 +47,7 @@ class TestStorageDiscoveryMgmtStorageDiscoveryWorkspacesOperations(AzureMgmtReco
                     "capabilities": {
                         "azureBlobStorage": {
                             "capacityDetails": {"status": "str"},
-                            "prefixDefinitions": [
+                            "prefixConfigurations": [
                                 {"containerName": "str", "storageAccountName": "str", "prefix": "str"}
                             ],
                         }
@@ -83,7 +83,7 @@ class TestStorageDiscoveryMgmtStorageDiscoveryWorkspacesOperations(AzureMgmtReco
                     "capabilities": {
                         "azureBlobStorage": {
                             "capacityDetails": {"status": "str"},
-                            "prefixDefinitions": [
+                            "prefixConfigurations": [
                                 {"containerName": "str", "prefix": "str", "storageAccountName": "str"}
                             ],
                         }